+++ backend/src/main/java/managers/concretes/CategoryManager.java
@@ -1,5 +1,6 @@
 package managers.concretes;
 
+import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -30,21 +31,21 @@ public class CategoryManager extends BasePostgreSqlManager<Category> implements 
 
 	@Override
 	public Result insert(Category category) {
-		try {
-			connect();
-		} catch (SQLException e) {
-			return new ErrorDataResult<>(new SQLException(DbErrorMessages.CONNECTION_FAILED, e).toString());
-		}
-
-		PreparedStatement statement;
-		try {
-			statement = connection.prepareStatement(SqlStrings.INSERT_STRING);
+		Result connectionResult = connect();
+		if(!connectionResult.isSuccess())
+			return connectionResult;
+		
+		try {	
+			PreparedStatement statement = connection.prepareStatement(SqlStrings.INSERT_STRING);
 			statement.setString(1, category.getCategoryName());
-		} catch (SQLException e) {
-			return new ErrorDataResult<>(new SQLException(DbErrorMessages.INSERTION_FAILED, e).toString());
+			
+			return super.insert(statement);
+		} catch (SQLException e1) {
+			e1.printStackTrace();
+			return new ErrorDataResult<>(DbErrorMessages.INSERTION_FAILED);
+		} finally {
+			disconnect();
 		}
-
-		return super.insert(statement);
 	}
 
 	@Override
@@ -55,17 +56,41 @@ public class CategoryManager extends BasePostgreSqlManager<Category> implements 
 
 	@Override
 	public Result delete(Long id) {
-		return super.deleteById(SqlStrings.DELETE_BY_ID_STRING, id);
+		try {
+			Result connectionResult = connect();
+			if(!connectionResult.isSuccess())
+				return connectionResult;
+			
+			return super.deleteById(SqlStrings.DELETE_BY_ID_STRING, id);
+		} finally {
+			disconnect();
+		}
 	}
 
 	@Override
 	public DataResult<Category> find(Long id) {
-		return super.find(SqlStrings.SELECT_BY_ID_STRING, id);
+		try {
+			Result connectionResult = connect();
+			if(!connectionResult.isSuccess())
+				return new ErrorDataResult<>(connectionResult.getMessage());
+			
+			return super.find(SqlStrings.SELECT_BY_ID_STRING, id);	
+		} finally {
+			disconnect();
+		}
 	}
 
 	@Override
 	public DataResult<List<Category>> listAll() {
-		return super.listAll(SqlStrings.SELECT_STRING);
+		try {
+			Result connectionResult = connect();
+			if(!connectionResult.isSuccess())
+				return new ErrorDataResult<>(connectionResult.getMessage());
+			
+			return super.listAll(SqlStrings.SELECT_STRING);
+		} finally {
+			disconnect();
+		}
 	}
 
 	private static class SqlStrings {
